pyproject.toml
[tool.poetry]
name = "xelt.py"
version = "3.0.0"
description = "Xelt is a multi-purpose discord bot with features ranging from user information to music!"
authors = ["No767, Edqe14, ThEditor"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4"
discord-py = {extras = ["voice"], version = "^2.2.2"}
python-dotenv = "^1.0.0"
aiodns = "^3.0.0"
brotli = "^1.0.9"
uvloop = "^0.17.0"
psutil = "^5.9.4"
prisma = "^0.8.2"
orjson = "^3.8.8"
ciso8601 = "^2.3.0"
anyio = "^3.6.2"
redis = {extras = ["hiredis"], version = "^4.5.2"}
faust-cchardet = "^2.1.18"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.21.0"
pyinstrument = "^4.4.0"
pyright = "^1.1.299"


[tool.poetry.group.tests.dependencies]
pytest = "^7.2.1"
pytest-asyncio = "^0.20.3"
pytest-cov = "^4.0.0"
nox = "^2022.11.21"
nox-poetry = "^1.0.2"


[tool.poetry.group.types.dependencies]
types-psutil = "^5.9.5.9"
types-redis = "^4.5.2.0"

[tool.isort]
profile = 'black'

[tool.pyright]
include = ["bot/**"]
exclude = ["**/__pycache__", "**/.mypy_cache", "**/.dmpypy.json"]
ignore = ["docker"]
reportMissingImports = true
reportMissingTypeStubs = false

[tool.bandit]
skips = ["B311", "B101"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
